# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: C5 360相机巡查标注系统 API
  version: 1.0.0
  license:
    name: Proprietary
  description: >
    Spec-first OpenAPI 3 contract for project C5 (360-camera inspection &
    annotation system). This single document is the authoritative source for
    oapi-codegen v2 → Go server stubs + TypeScript + Kotlin/Retrofit clients.


    Conventions:

    - Every response uses the shared envelope `{success, data, error, meta?}`.
      List endpoints populate `meta` with `{total, page, page_size}`.

    - All IDs are int64 (BIGSERIAL in the DB). `client_uuid` / `job_uuid` are
      UUID strings (offline idempotency / public job handles).

    - All geometry on the wire is WGS84 / SRID 4326 GeoJSON. GCJ-02 is NEVER
      persisted or transmitted; it is a client-only render-boundary concern.

    - All timestamps are RFC3339 / ISO-8601 UTC (`format: date-time`).

    - Error codes are a frozen, canonical enum (see ErrorObject.code).


    Out-of-band (NOT expressible in OpenAPI):

    - `GET /exports/{job_uuid}/events` is a Server-Sent Events (`text/event-stream`)
      endpoint emitting `event: progress` and a terminal `event: done`. It is
      hand-written with `GET /exports/{job_uuid}` as the polling fallback
      (Tencent CLB/CDN may buffer SSE). It is intentionally absent from `paths`.

    - Dictionary / config pull uses ETag (`If-None-Match: "<content_hash>"` →
      304 Not Modified when unchanged). 304 responses carry no body.

...
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [users]
      operationId: listUsers
      summary: List users
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: q
          in: query
          description: Search by username / display_name
          schema: { type: string }
        - name: is_active
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [users]
      operationId: createUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [users]
      operationId: getUser
      summary: Get a user by id
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [users]
      operationId: updateUser
      summary: Update a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [users]
      operationId: deleteUser
      summary: Soft-delete a user
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{id}/roles:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [users]
      operationId: getUserRoles
      summary: Get a user's roles (casbin g-rules)
      responses:
        '200':
          description: User roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [users]
      operationId: setUserRoles
      summary: Replace a user's roles (casbin g-rules)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUserRolesRequest'
      responses:
        '200':
          description: Roles updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{id}/password:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      tags: [users]
      operationId: resetUserPassword
      summary: Admin reset a user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # Roles
  # ===========================================================================
  /roles:
    get:
      tags: [roles]
      operationId: listRoles
      summary: List roles
      responses:
        '200':
          description: Role list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [roles]
      operationId: createRole
      summary: Create a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreate'
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /roles/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      tags: [roles]
      operationId: updateRole
      summary: Update a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdate'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [roles]
      operationId: deleteRole
      summary: Delete a role (system roles protected)
      responses:
        '200':
          description: Role deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /roles/{id}/permissions:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [roles]
      operationId: getRolePermissions
      summary: Get a role's permissions (casbin p-rules)
      responses:
        '200':
          description: Role permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionListResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [roles]
      operationId: setRolePermissions
      summary: Replace a role's permissions (casbin p-rules)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetRolePermissionsRequest'
      responses:
        '200':
          description: Permissions updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionListResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # Permissions
  # ===========================================================================
  /permissions:
    get:
      tags: [permissions]
      operationId: listPermissions
      summary: Full permission catalog (UI tree)
      responses:
        '200':
          description: Permission catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionListResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # Dictionary
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access JWT (golang-jwt/v5). Refresh is an opaque Redis token.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
  responses:
...
    Unauthenticated:
      description: Missing/invalid/expired bearer token (UNAUTHENTICATED | TOKEN_EXPIRED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: casbin (sub,obj,act) denied (FORBIDDEN)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found or soft-deleted (NOT_FOUND)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Unique conflict / RESTRICT / illegal state transition (CONFLICT)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationFailed:
      description: Boundary schema / business validation failed (VALIDATION_FAILED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unhandled error (INTERNAL); never leaks stack/SQL
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Meta:
      type: object
      properties:
        total: { type: integer, format: int64 }
        page: { type: integer }
        page_size: { type: integer }
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - TOKEN_EXPIRED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - MEDIA_VERIFY_FAILED
        - DICT_VERSION_RETIRED
        - RATE_LIMITED
        - INTERNAL
    ErrorDetail:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message: { type: string }
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: false }
        data:
          type: object
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    EmptyResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'

    # -------------------------------------------------------------------------
    # GeoJSON primitives (WGS84 / 4326)
    # -------------------------------------------------------------------------
...
    User:
      type: object
      required: [id, username, display_name, is_active, created_at, updated_at]
      description: password_hash is NEVER exposed.
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        display_name: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        avatar_media_id: { type: integer, format: int64, nullable: true }
        is_active: { type: boolean }
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Role:
      type: object
      required: [id, code, name, is_system, sort_order]
      properties:
        id: { type: integer, format: int64 }
        code: { type: string }
        name: { type: string }
        description: { type: string }
        is_system: { type: boolean }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Permission:
      type: object
      required: [id, code, name, object, action]
      properties:
        id: { type: integer, format: int64 }
        code: { type: string }
        name: { type: string }
        object: { type: string }
        action: { type: string }
        group_name: { type: string }
        description: { type: string }
        sort_order: { type: integer }
...
    ResetPasswordRequest:
      type: object
      required: [new_password]
      properties:
        new_password: { type: string, format: password }
    UserCreate:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
        display_name: { type: string }
        phone: { type: string }
        email: { type: string }
        is_active: { type: boolean, default: true }
        role_ids:
          type: array
          items: { type: integer, format: int64 }
    UserUpdate:
      type: object
      properties:
        display_name: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        avatar_media_id: { type: integer, format: int64, nullable: true }
        is_active: { type: boolean }
    SetUserRolesRequest:
      type: object
      required: [role_ids]
      properties:
        role_ids:
          type: array
          items: { type: integer, format: int64 }
    RoleCreate:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: string }
        sort_order: { type: integer }
    RoleUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        sort_order: { type: integer }
    SetRolePermissionsRequest:
      type: object
      required: [permission_ids]
      properties:
        permission_ids:
          type: array
          items: { type: integer, format: int64 }
...
    UserResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/User'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    UserListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    RoleResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/Role'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    RoleListResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    PermissionListResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...